+++ processor/src/test/resources/bad_input/com/example/component/NonObservableWithDisposeOnDeactivateModel.java
@@ -1,0 +1,22 @@
+package com.example.component;
+
+import arez.annotations.ArezComponent;
+import arez.annotations.Observable;
+
+@ArezComponent( observable = false, disposeOnDeactivate = true )
+public abstract class NonObservableWithDisposeOnDeactivateModel
+{
+  private long _field;
+
+  @Observable
+  public long getField()
+  {
+    return _field;
+  }
+
+  @Observable
+  public void setField( final long field )
+  {
+    _field = field;
+  }
+}